requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dnspython==2.2.1
future==0.18.2
h11==0.13.0
idna==3.3
numpy==1.23.3
outcome==1.2.0
pandas==1.5.0
pdfrw==0.4
plotly==5.10.0
psutil==5.9.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tenacity==8.1.0
trio==0.21.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.11
websockets==10.3
wsproto==1.1.0
